add protected route component
