+++ src/App.tsx
@@ -8,6 +8,7 @@ import { Landing } from "./Landing";
 import Callback from "./Callback";
 import Logout from "./Logout";
 import Profile from "./Profile";
+import { ProtectedRoute } from "./components/ProtectedRoute";
 
 function App() {
   return (
@@ -16,7 +17,13 @@ function App() {
         <Route path="/" element={<Landing />}></Route>
         <Route path="/oauth_callback" element={<Callback />}></Route>
         <Route path="/logout" element={<Logout />}></Route>
-        <Route path="/dashboard" element={<Dashboard />}>
+        <Route
+          path="/dashboard"
+          element={
+            <ProtectedRoute>
+              <Dashboard />
+            </ProtectedRoute>
+          }>
           <Route path="/dashboard/profile" element={<Profile />}></Route>
         </Route>
       </Routes>
